Using a float on the progress on app updates
Not sure if needed but I saw some funny behavior with the progress bar

+++ app/src/main/java/eu/kanade/tachiyomi/data/updater/UpdaterService.kt
@@ -21,9 +21,9 @@ import eu.kanade.tachiyomi.util.system.isServiceRunning
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
-import java.io.File
 import timber.log.Timber
 import uy.kohesive.injekt.injectLazy
+import java.io.File
 
 class UpdaterService : Service() {
 
@@ -100,7 +100,7 @@ class UpdaterService : Service() {
             var lastTick = 0L
 
             override fun update(bytesRead: Long, contentLength: Long, done: Boolean) {
-                val progress = (100 * bytesRead / contentLength).toInt()
+                val progress = (100 * (bytesRead.toFloat() / contentLength)).toInt()
                 val currentTime = System.currentTimeMillis()
                 if (progress > savedProgress && currentTime - 200 > lastTick) {
                     savedProgress = progress
